# api_spec.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/status": {
            "get": {
                "operationId": "getStatus",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "operationId": "findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid or expired user token"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/users/@me": {
            "get": {
                "operationId": "findMyProfile",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserEntity"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid or expired user token"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "operationId": "findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid or expired user token"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/auth/github": {
            "get": {
                "operationId": "githubLogin",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/auth/github/callback": {
            "get": {
                "operationId": "githubCallback",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/guilds": {
            "post": {
                "operationId": "create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateGuildDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GuildEntity"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid or expired user token"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "get": {
                "operationId": "findAll",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "The page to retrieve. The default is 0.",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "The number of results to return. By default, 10 results are returned per page.",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "401": {
                        "description": "Invalid or expired user token"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/guilds/{slug}": {
            "get": {
                "operationId": "findOne",
                "parameters": [
                    {
                        "name": "slug",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GuildEntity"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid or expired user token"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/guilds/availability/{slug}": {
            "get": {
                "operationId": "checkSlugAvailability",
                "parameters": [
                    {
                        "name": "slug",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "401": {
                        "description": "Invalid or expired user token"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/guilds/{id}": {
            "patch": {
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGuildDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid or expired user token"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "operationId": "remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid or expired user token"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/guilds/{guildId}/invites": {
            "get": {
                "operationId": "findAll",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/guilds/invites/{id}": {
            "get": {
                "operationId": "findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/guilds/{guildId}/members/{userId}": {
            "get": {
                "operationId": "findOne",
                "parameters": [
                    {
                        "name": "guildId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "info": {
        "title": "Mescla Backend",
        "description": "The Mescla Backend",
        "version": "0.0.1",
        "contact": {}
    },
    "tags": [],
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "UserEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "username": {
                        "type": "string"
                    },
                    "displayName": {
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "password": {
                        "type": "string",
                        "nullable": true
                    },
                    "avatarUrl": {
                        "type": "string",
                        "nullable": true
                    },
                    "authProvider": {
                        "type": "object"
                    },
                    "thirdPartyId": {
                        "type": "string",
                        "nullable": true
                    },
                    "lastLoginAt": {
                        "format": "date-time",
                        "type": "string",
                        "nullable": true
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "deletedAt": {
                        "format": "date-time",
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": [
                    "id",
                    "username",
                    "displayName",
                    "email",
                    "password",
                    "avatarUrl",
                    "authProvider",
                    "thirdPartyId",
                    "lastLoginAt",
                    "createdAt",
                    "updatedAt",
                    "deletedAt"
                ]
            },
            "CreateGuildDto": {
                "type": "object",
                "properties": {
                    "slug": {
                        "type": "string",
                        "maxLength": 28
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": [
                    "slug",
                    "name",
                    "description"
                ]
            },
            "GuildEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "deletedAt": {
                        "format": "date-time",
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": [
                    "id",
                    "slug",
                    "name",
                    "description",
                    "createdAt",
                    "updatedAt",
                    "deletedAt"
                ]
            },
            "UpdateGuildDto": {
                "type": "object",
                "properties": {
                    "slug": {
                        "type": "string",
                        "maxLength": 28
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            }
        }
    }
}